Refactor verify_session to do the right thing
verify_session was previously not doing what I intended it to do, but
was working like intended none the less.  This change aims to wait a
number of seconds making sure the session still exists.  After this time
period we deem the session valid, if at any point when waiting to
validate the session it doesn't exist we try to create a new one.