Move code to obtain session to class instance
This commit changes the behavior of the class in that storage of the
lock is not handled in the class instance and not in an instance
variable at the class level.  This change should be mostly non-breaking.

The interaction with the code should remain mostly the same, you don't
have to call `start` before calling `lock`.  It behaves much closer to
the way `File` does in that `File.open`(without a block) returns an
instance a a class that you can do things with.  This now works in the
same way, from `lock` we return an instance of `Semaphore` so that you
can do what you need to do.

In addition to the above change, `renew` was extracted to make #22
easier to implement.

Fixes #23